+++ .idea/encodings.xml
@@ -1,11 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Encoding" useUTFGuessing="true" native2AsciiForPropertiesFiles="false">
-    <file url="file://$PROJECT_DIR$/src/controllers/IndexController.php" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/models/Articles.php" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/views/Template/editArticleForm.php" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/views/articlePage.php" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/views/editArticlePage.php" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/views/searchPage.php" charset="UTF-8" />
-  </component>
+  <component name="Encoding" useUTFGuessing="true" native2AsciiForPropertiesFiles="false" />
 </project>
+
