--- articlesDesk.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace pers1307\blog;
-
-use pers1307\blog\controllers;
-
-require __DIR__ . '/vendor/autoload.php';
-
-//$articlesDeskContoller = new controllers\ArticlesDeskController();
-//$articlesDeskContoller->controlContent();
