+++ composer.json
@@ -10,8 +10,7 @@
     ],
     "require": {
         "ko-ko-ko/assert": "0.2.x",
-        "phroute/phroute": "2.0.0",
-        "vlucas/valitron": "~1.2"
+        "phroute/phroute": "2.0.0"
     },
 	"autoload": {
 		"psr-4": {
