+++ src/autorization/Autorization.php
@@ -59,8 +59,7 @@ class Autorization
         $user = $users->findByCreditionals($login);
 
         if ($user !== null) {
-            //if (!\password_verify($password, $user->getPassword())) {
-            if ($password !== $user->getPassword()) {
+            if (!\password_verify($password, $user->getPassword())) {
                 return false;
             } else {
                 return true;
