+++ src/config/Config.php
@@ -12,8 +12,4 @@ namespace pers1307\blog\config;
 
 class Config extends DefaultConfig
 {
-    const HTTP_HOST    = 'localhost';
-    const PDO_DSN      = 'mysql:dbname=myblog;host=127.0.0.1';
-    const PDO_USER     = 'root';
-    const PDO_PASSWORD = '';
 }
