--- test2.php
@@ -1,5 +1,0 @@
-<?php
-
-$result = $_POST['delete'];
-
-echo $result;
