--- test3.php
@@ -1,9 +1,0 @@
-<?php
-
-$login = $_POST['login'];
-$password = $_POST['password'];
-
-$result['login'] = $login;
-$result['password'] = $password;
-
-echo json_encode($result);
